Populate encoder logic

+++ src/bvc_common/spiht/spiht_config.h
@@ -3,5 +3,5 @@
 class bvc_spiht_config
 {
 public:
-	float bpp;
+	double bpp;
 };
